requirements.txt
pytorch==1.3.1
torchvision
opencv==3.4.2
tqdm
aiogram
aiohttp
imageio==2.9.0
imageio-ffmpeg==0.4.2
cmake
dlib
aioredis
